\name{as.character.OpenGammaErrorValue}
\alias{as.character,OpenGammaErrorValue-method}
\title{Get the text description of an error}
\description{
	Returns a string describing the error.
}
\arguments{
	\item{x}{An OpenGammaErrorValue object}
	\item{...}{Ignored}
}
\usage{
	\S4method{as.character}{OpenGammaErrorValue}(x, ...)
}
