\name{as.data.frame.VolalitySurfaceSnapshot}
\alias{as.data.frame.VolatilitySurfaceSnapshot}
\alias{as.data.frame,VolatilitySurfaceSnapshot-method}
\title{Get a data frame representation of a voltility surface snapshot}
\description{
	Returns a data frame representation.
}
\arguments{
	\item{x}{A VolatilitySurfaceSnapshot object}
	\item{row.names}{Ignored}
	\item{optional}{Ignored}
	\item{...}{Ignored}
}
\usage{
	\S4method{as.data.frame}{VolatilitySurfaceSnapshot}(x, row.names, optional, ...)
}
