\name{as.data.frame.YieldCurveSnapshot}
\alias{as.data.frame.YieldCurveSnapshot}
\alias{as.data.frame,YieldCurveSnapshot-method}
\title{Get a data frame representation of a voltility surface snapshot}
\description{
	Returns a data frame representation.
}
\arguments{
	\item{x}{A YieldCurveSnapshot object}
	\item{row.names}{Ignored}
	\item{optional}{Ignored}
	\item{...}{Ignored}
}
\usage{
	\S4method{as.data.frame}{YieldCurveSnapshot}(x, row.names, optional, ...)
}
