\name{classNames.FudgeMsg}%Misc
\alias{classNames.FudgeMsg}
\title{Get the class names of an encoded object}
\description{
	Returns the fields with ordinal 0. When an object is serialized to a Fudge encoding, class names
	may be written with field ordinal 0 to specify a given sub-class.
}
\arguments{
	\item{msg}{A Fudge message object}
}
