\name{column.ViewComputationResultModel}
\alias{column.ViewComputationResultModel}
\title{Find column from a data result}
\description{
	Finds a column with a given value specification from a named configuration
	within a result object. A list is returned which the column values and
	labels corresponding to the computation target identifiers from each value.
}
\arguments{
	\item{data}{Result object to process}
	\item{config}{Name of the configuration to process}
	\item{col}{Value specification to extract - e.g. as returned from columns.ViewComputationResultModel}
}
