\name{columns.ViewComputationResultModel}
\alias{columns.ViewComputationResultModel}
\title{Find a set of satisfying columns from a data frame}
\description{
	Finds the column names from a data frame that match the named value and can
	satisfy any constraints on the requirement.
}
\arguments{
	\item{data}{Data frame to search the columns of}
	\item{valueRequirement}{Value requirement string to match}
}
