\name{isEmpty.FudgeMsg}%Misc
\alias{isEmpty.FudgeMsg}
\title{Test for an empty Fudge messsage}
\description{
	Tests whether a Fudge message is empty - that is, it contains no fields.
	Returns \code{true} if the message has no fields, \code{false} otherwise.
}
\arguments{
	\item{x}{A Fudge message object}
}
