\name{name.ValueRequirement}%Value
\alias{name.ValueRequirement}
\alias{name.ValueSpecification}
\title{Extract the name component of a value requirement/specification string}
\description{
	Parses a value requirement/specification string and returns the name element.
}
\arguments{
	\item{requirement}{Requirement string to parse}
}
