\name{new.ValueRequirement}
\alias{new.ValueRequirement}
\alias{new.ValueSpecification}
\title{Create a new value requirement/specification string}
\description{
	Creates a requirement/specification string from a name and properties pair. Note that
	the string representations of requirements and specifications are the same.
}
\arguments{
	\item{name}{Value requirement name}
	\item{properties}{Value properties}
}
