\name{numFields.FudgeMsg}%Misc
\alias{numFields.FudgeMsg}
\title{Count the fields in a Fudge messsage}
\description{
	Fetches the number of fields in a Fudge message. If only a count is required
	this is more efficient than calling \code{length} on the list returned by
	\code{fields.FudgeMsg}.
}
\arguments{
	\item{x}{A Fudge message object}
}
