\name{object.FudgeMsg}
\alias{object.FudgeMsg}
\title{Declare an object based on a Fudge message}
\description{
	Declares a class template that holds a fudge message and can return it through the
	\code{toFudgeMsg} function. Additional functions should be declared on the class
	so that it can be queried or manipulated in R.
}
\arguments{
	\item{className}{Name of the class to declare}
}
