\name{properties.ValueRequirement}%Value
\alias{properties.ValueRequirement}
\alias{properties.ValueSpecification}
\title{Extract the properties component of a value requirement/specification string}
\description{
	Parses a value requirement/specification string and returns the properties element.
}
\arguments{
	\item{requirement}{Requirement string to parse}
}
