\name{satisfiedBy.ValueProperties}
\alias{satisfiedBy.ValueProperties}
\title{Test the satisfied-by relationship between two value property sets}
\description{
	Tests if one set of value properties can satisfy the contraints described by another.
	Returns TRUE if A is satisfied by B (or equally worded as if B satisfies A) and FALSE
	otherwise. A and B may be either string representations of value properties or
	instances of the ValueProperties object type.
}
\arguments{
	\item{a}{Constraints to test satisfaction of}
	\item{b}{Properties to statisfaction of the constraints with}
}
