\name{values.FudgeMsg}%Misc
\alias{values.FudgeMsg}
\title{Get values from a Fudge messsage}
\description{
	Fetches all of the values from a Fudge message. If the field names or ordinals
	are not required, this may be more efficient than calling fields.FudgeMsg
}
\arguments{
	\item{x}{A Fudge message object}
}
