\name{object.ExternalRef}
\alias{object.ExternalRef}
\title{Declare an object based on a tracked reference}
\description{
	Declares a class template that holds a tracked reference and can return it through the
	\code{toString} function.
}
\arguments{
	\item{className}{Name of the class to declare}
}
